// frontend/src/api/generated/system/system.ts
export type getSystemStatusApiV1SystemStatusGetResponse200 = {
  data: SystemStatusResponse;
  status: 200;
};

export type getSystemStatusApiV1SystemStatusGetResponseSuccess =
  getSystemStatusApiV1SystemStatusGetResponse200 & {
    headers: Headers;
  };
export type getSystemStatusApiV1SystemStatusGetResponse =
  getSystemStatusApiV1SystemStatusGetResponseSuccess;